requirements.txt
django==3.2.3
djangorestframework==3.12.4
django-rest-auth[with_social]==0.9.5
psycopg2==2.8.6
